# docs/evidence/contracts/iskai-openapi.current.json
{
  "components": {
    "schemas": {
...
      "CatalogEntityCreate": {
        "properties": {
          "code": {
            "maxLength": 24,
            "minLength": 2,
            "pattern": "^[A-Z0-9-]+$",
            "title": "Code",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Parent Id",
            "type": "string"
          }
        },
        "required": [
          "parent_id",
          "code",
          "name"
        ],
        "title": "CatalogEntityCreate",
        "type": "object"
      },
      "CatalogEntityUpdate": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CatalogEntityUpdate",
        "type": "object"
      },
...
      "DepartmentCreate": {
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "track_label": {
            "default": "Academic track",
            "maxLength": 50,
            "minLength": 1,
            "title": "Track Label",
            "type": "string"
          },
          "tracks": {
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "DepartmentCreate",
        "type": "object"
      },
      "DepartmentOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "track_label": {
            "title": "Track Label",
            "type": "string"
          },
          "tracks": {
            "items": {
              "type": "string"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "track_label",
          "tracks",
          "created_at"
        ],
        "title": "DepartmentOut",
        "type": "object"
      },
      "DepartmentUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "track_label": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Label"
          },
          "tracks": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracks"
          }
        },
        "title": "DepartmentUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Centralized AI-Powered Thesis Library using Retrieval-Augmented Generation for the College of Computing Studies, Information and Communication Technology, Isabela State University, Echague.",
    "title": "ISU Thesis AI Library API",
    "version": "2.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/catalog/programs": {
      "post": {
        "operationId": "create_program_catalog_programs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Program",
        "tags": [
          "Academic catalog"
        ]
      }
    },
    "/catalog/programs/{entity_id}": {
      "patch": {
        "operationId": "update_program_catalog_programs__entity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Program",
        "tags": [
          "Academic catalog"
        ]
      }
    },
    "/catalog/specializations": {
      "post": {
        "operationId": "create_specialization_catalog_specializations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Specialization",
        "tags": [
          "Academic catalog"
        ]
      }
    },
    "/catalog/specializations/{entity_id}": {
      "patch": {
        "operationId": "update_specialization_catalog_specializations__entity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Specialization",
        "tags": [
          "Academic catalog"
        ]
      }
    },
...
    "/departments/": {
      "get": {
        "description": "Fetch departments for server-validated filtering.",
        "operationId": "list_departments_departments__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DepartmentOut"
                  },
                  "title": "Response List Departments Departments  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Departments",
        "tags": [
          "Departments"
        ]
      },
      "post": {
        "description": "Create a department and its tracks.",
        "operationId": "create_department_departments__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The request payload or parameters were rejected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Department",
        "tags": [
          "Departments"
        ]
      }
    },
    "/departments/{department_id}": {
      "delete": {
        "description": "Delete a department.",
        "operationId": "delete_department_departments__department_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "409": {
            "description": "The request conflicts with the current state of the resource."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Department",
        "tags": [
          "Departments"
        ]
      },
      "put": {
        "description": "Update a department.",
        "operationId": "update_department_departments__department_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The request payload or parameters were rejected."
          },
          "404": {
            "description": "The resource does not exist, or is not visible to this account."
          },
          "409": {
            "description": "The request conflicts with the current state of the resource."
          },
          "422": {
            "description": "The request was well formed but failed validation."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Department",
        "tags": [
          "Departments"
        ]
      }
    },
...
  }
}